refactor : LottosGenerate의 default 메서드를 통해 generate 처리
