+++ src/test/java/lotto/RanksTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.domain.Lottos;
-import lotto.domain.ManualLottos;
+import lotto.domain.LottosMaker;
 import lotto.domain.Ranks;
 import org.junit.jupiter.api.Test;
 
@@ -22,7 +22,7 @@ public class RanksTest {
         ranksInit.add(Rank.FIRST);
         int[] winLotto = new int[] {1, 2, 3, 4, 5, 6};
         List<String> manualLotto = Arrays.asList("1, 2, 3, 4, 5, 8");
-        Lottos lottos = new Lottos(new ManualLottos(manualLotto).getManualLottos());
+        Lottos lottos = new LottosMaker().generate(manualLotto);
         Ranks ranks = new Ranks(winLotto, 8, lottos);
         assertThat(ranks.getRanks()).contains(Rank.SECOND);
     }
